Phase 6-7: Move vLLM engine and KV transfer backends
- Move deployments/llm/vllm/vllm_engine.py → engines/vllm/vllm_engine.py
- Move deployments/llm/vllm/vllm_models.py → engines/vllm/vllm_models.py
- Move deployments/llm/vllm/kv_transfer_backends → engines/vllm/kv_transfer
- Rename lmcache_connector_v1.py → lmcache.py
- Rename nixl_connector.py → nixl.py
- Update all imports
