+++ python/ray/llm/_internal/serve/engines/vllm/kv_transfer/nixl.py
@@ -1,6 +1,6 @@
 import os
 
-from ray.llm._internal.serve.deployments.llm.vllm.kv_transfer_backends.base import (
+from ray.llm._internal.serve.engines.vllm.kv_transfer.base import (
     BaseConnectorBackend,
 )
