+++ python/ray/llm/tests/serve/cpu/deployments/llm/vllm/kv_transfer_backends/test_nixl_connector.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
-from ray.llm._internal.serve.deployments.llm.vllm.kv_transfer_backends.nixl_connector import (
+from ray.llm._internal.serve.engines.vllm.kv_transfer.nixl_connector import (
     NixlConnectorBackend,
 )
 from ray.serve.llm import LLMConfig
